Use python 3(.7) and relevant pip. (#72)
* Switch to alpine and reduce heap size.

Switch to alpine jdk image so as to reduce docker image size, and reduce heap size (for our expected usage).

* Delete deploy.sh

Not needed for this.

* Change from alpine to base openjdk.

It appears Openjdk doesn't support alpine anymore, and so pulling `openjdk:11-alpine` fails. This will result in a larger image.

* Change Minecraft to 1.15.2

Change minecraft to 1.15.2, the latest version supported by paper.

* Update to 1.16.2

Although upstream is 1.16.1 at most, 1.16.2 jar exists so should be ok to use.

* Fix bug with get-pip. Use latest version of paper.

Early 2021 (post python 2.x deprecation), the line `RUN python get-pip.py` stopped working; this is due to the URL now loading the 3.6+ version of get-pip, involving an f-string.
The ideal fix would be to use python3 in the run container, as 2.x is now EOL; however this commit fixes it by using the older version of get-pip.

* Even more fixes for pip.

Seems to be the way of the dodo for 2.7, as it should be. This should really get fixed in upstream; I have an open PR.

* Implement change to python3.
Change maintainer label.

* Reset rest of dockerfile to base.
Now I understand args and envs, so we don't need these changes.

Co-authored-by: fisher <fisher@theeu.uk>